+++ experimental/packages/opentelemetry-sdk-metrics/test/export/TestMetricProducer.ts
@@ -14,10 +14,17 @@
  * limitations under the License.
  */
 
-/**
- * AggregationTemporality indicates the way additive quantities are expressed.
- */
-export enum AggregationTemporality {
-  DELTA,
-  CUMULATIVE,
+import { CollectionResult } from '../../src/export/MetricData';
+import { MetricProducer } from '../../src/export/MetricProducer';
+import { defaultResource } from '../util';
+
+export const emptyResourceMetrics = { resource: defaultResource, scopeMetrics: [] };
+
+export class TestMetricProducer implements MetricProducer {
+  async collect(): Promise<CollectionResult> {
+    return {
+      resourceMetrics: { resource: defaultResource, scopeMetrics: [] },
+      errors: [],
+    };
+  }
 }