+++ experimental/packages/opentelemetry-sdk-node/src/sdk.ts
@@ -21,6 +21,7 @@ import {
   registerInstrumentations
 } from '@opentelemetry/instrumentation';
 import {
+  Detector,
   detectResources,
   envDetector,
   processDetector,
@@ -62,6 +63,7 @@ export class NodeSDK {
   private _instrumentations: InstrumentationOption[];
 
   private _resource: Resource;
+  private _resourceDetectors: Detector[];
 
   private _autoDetectResources: boolean;
 
@@ -74,6 +76,7 @@ export class NodeSDK {
    */
   public constructor(configuration: Partial<NodeSDKConfiguration> = {}) {
     this._resource = configuration.resource ?? new Resource({});
+    this._resourceDetectors = configuration.resourceDetectors ?? [envDetector, processDetector];
 
     this._serviceName = configuration.serviceName;
 
@@ -166,12 +169,9 @@ export class NodeSDK {
   }
 
   /** Detect resource attributes */
-  public async detectResources(
-    config?: ResourceDetectionConfig
-  ): Promise<void> {
+  public async detectResources(): Promise<void> {
     const internalConfig: ResourceDetectionConfig = {
-      detectors: [envDetector, processDetector],
-      ...config,
+      detectors: this._resourceDetectors,
     };
 
     this.addResource(await detectResources(internalConfig));