+++ experimental/packages/opentelemetry-sdk-node/src/types.ts
@@ -17,7 +17,7 @@
 import type { ContextManager, SpanAttributes } from '@opentelemetry/api';
 import { TextMapPropagator } from '@opentelemetry/api';
 import { InstrumentationOption } from '@opentelemetry/instrumentation';
-import { Resource } from '@opentelemetry/resources';
+import { Detector, Resource } from '@opentelemetry/resources';
 import { MetricReader, View } from '@opentelemetry/sdk-metrics';
 import {
   Sampler,
@@ -35,6 +35,7 @@ export interface NodeSDKConfiguration {
   views: View[]
   instrumentations: InstrumentationOption[];
   resource: Resource;
+  resourceDetectors: Detector[];
   sampler: Sampler;
   serviceName?: string;
   spanProcessor: SpanProcessor;